Change writing M3U filenames so that filenames beginning with a hash work
Since the M3U file attaches a special meaning to lines beginning with a
hash, any filenames beginning with a hash (might be rare) will not
appear when the playlist is loaded. This trick of appended the current
directory './' to the beginning of the line seems to work fine, at least
with Audacious.